refactor(data-table): consolidate stories into a single file and enhance functionality with new features and tests
