--- packages/nimbus/src/components/data-table/stories/custom-column.tsx
@@ -1,12 +1,0 @@
-import type { StoryObj } from "@storybook/react-vite";
-
-import { columns, rows } from "../test-data";
-import type { DataTableProps } from "../data-table.types";
-import { DataTableWithModals } from "./utils";
-
-type Story = StoryObj<DataTableProps>;
-
-export const CustomColumn: Story = {
-  render: (args) => <DataTableWithModals {...args} onRowClick={() => {}} />,
-  args: { columns, rows },
-};
