--- packages/nimbus/src/components/data-table/stories/multiline-headers.tsx
@@ -1,17 +1,0 @@
-import type { StoryObj } from "@storybook/react-vite";
-import { multilineHeadersColumns, multilineHeadersData } from "../test-data";
-import type { DataTableProps } from "../data-table.types";
-import { DataTableWithModals } from "./utils";
-
-type Story = StoryObj<DataTableProps>;
-
-export const MultilineHeaders: Story = {
-  render: (args) => <DataTableWithModals {...args} />,
-  args: {
-    columns: multilineHeadersColumns,
-    rows: multilineHeadersData,
-    allowsSorting: true,
-    isResizable: true,
-    onRowClick: () => {},
-  },
-};
